Just use a single associative array.
Thought it would be easier to use "${(P)}" with `typeset` arrays, but it
is not at all easier and this change reduces the line count.

Also, introduce a path structure that can be used to create ids in DOM.
Identifiers can be snake_case, but not kabob-case, and that should come
naturally in Zsh.